Extract tests for keys to own file
